refactor: #696 interview_turn をターン処理境界で再分割する
